requirements.txt
feedparser>=6.0.10
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.3.0
python-dateutil>=2.8.2
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
tqdm>=4.62.0
markdown>=3.4.0
gspread>=5.12.0
google-auth>=2.23.0
pytest>=8.2.0
pytest-cov>=4.1.0
